# swagger.yaml
swagger: '2.0'
info:
  version: v3.0
  title: rocket-skates
  description: Protocol for Bare Metal Provider
host: 'localhost:8092'
basePath: /api/v3
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  rebar:
    type: basic
paths:
...
  /bootenvs:
    get:
      operationId: list-bootenvs
      summary: List Bootenvs
      tags:
        - Bootenvs
      responses:
        '200':
          description: ''
          schema:
            type: array
            items:
              $ref: '#/definitions/bootenv-input'
        '401':
          $ref: '#/responses/trait:commonGetResponses:401'
        '404':
          $ref: '#/responses/trait:commonGetResponses:404'
        '500':
          $ref: '#/responses/trait:commonGetResponses:500'
  /bootenv:
    post:
      operationId: post-bootenv
      summary: Create Bootenv
      tags:
        - Bootenvs
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/bootenv-input'
            example:
              Available: false
              BootParams: ''
              Errors:
                - 'bootenv: windows-2012r2-install: missing kernel wimboot (/tftpboot/windows-2012r2/install/wimboot)'
                - 'bootenv: windows-2012r2-install: missing initrd boot/bcd (/tftpboot/windows-2012r2/install/boot/bcd)'
                - 'bootenv: windows-2012r2-install: missing initrd boot/boot.sdi (/tftpboot/windows-2012r2/install/boot/boot.sdi)'
                - 'bootenv: windows-2012r2-install: missing initrd rebar-winpe.wim (/tftpboot/windows-2012r2/install/rebar-winpe.wim)'
              Initrds:
                - boot/bcd
                - boot/boot.sdi
                - rebar-winpe.wim
              Kernel: wimboot
              Name: windows-2012r2-install
              OS:
                Codename: ''
                Family: ''
                Files: null
                IsoFile: rebar-win2012r2.iso
                IsoSha256: 8ce9f9b23aaf397913d927c7fda3343d411f6846ef7b94f29b51108e95c46f6e
                IsoUrl: ''
                Name: windows-2012r2
                Version: ''
              RequiredParams:
                - logging_servers
                - ntp_servers
                - operating-system-disk
                - provisioner-default-password-hash
                - proxy-servers
                - use-proxy
                - rebar-access_keys
                - rebar-machine_key
                - operating-system-license-key
                - operating-system-install-flavor
              Templates:
                - Name: pxelinux
                  Path: 'pxelinux.cfg/{{.Machine.HexAddress}}'
                  UUID: windows.pxelinux.tmpl
                - Name: ipxe
                  Path: '{{.Machine.Address}}.ipxe'
                  UUID: windows.ipxe.tmpl
                - Name: start-install
                  Path: '{{.Machine.Path}}/stage1.ps1'
                  UUID: stage1.cmd.tmpl
                - Name: unattend.xml
                  Path: '{{.Machine.Path}}/unattend.xml'
                  UUID: 2012r2-unattend.xml.tmpl
              TenantId: 1
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/bootenv-input'
        '401':
          $ref: '#/responses/trait:commonPostResponses:401'
        '409':
          $ref: '#/responses/trait:commonPostResponses:409'
        '500':
          $ref: '#/responses/trait:commonPostResponses:500'
  '/bootenvs/{name}':
    parameters:
      - name: name
        in: path
        required: true
        type: string
    get:
      operationId: get-bootenv
      summary: Get Bootenv
      tags:
        - Bootenvs
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/bootenv-input'
        '401':
          $ref: '#/responses/trait:commonGetResponses:401'
        '404':
          $ref: '#/responses/trait:commonGetResponses:404'
        '500':
          $ref: '#/responses/trait:commonGetResponses:500'
    patch:
      operationId: patch-bootenv
      summary: Patch Bootenv
      tags:
        - Bootenvs
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/patchinput'
      responses:
        '202':
          description: ''
          schema:
            $ref: '#/definitions/bootenv-input'
        '401':
          $ref: '#/responses/trait:commonPatchResponses:401'
        '404':
          $ref: '#/responses/trait:commonPatchResponses:404'
        '417':
          $ref: '#/responses/trait:commonPatchResponses:417'
        '500':
          $ref: '#/responses/trait:commonPatchResponses:500'
    put:
      operationId: put-bootenv
      summary: Update Bootenv
      tags:
        - Bootenvs
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/bootenv-input'
            example:
              Available: false
              BootParams: ''
              Errors:
                - 'bootenv: windows-2012r2-install: missing kernel wimboot (/tftpboot/windows-2012r2/install/wimboot)'
                - 'bootenv: windows-2012r2-install: missing initrd boot/bcd (/tftpboot/windows-2012r2/install/boot/bcd)'
                - 'bootenv: windows-2012r2-install: missing initrd boot/boot.sdi (/tftpboot/windows-2012r2/install/boot/boot.sdi)'
                - 'bootenv: windows-2012r2-install: missing initrd rebar-winpe.wim (/tftpboot/windows-2012r2/install/rebar-winpe.wim)'
              Initrds:
                - boot/bcd
                - boot/boot.sdi
                - rebar-winpe.wim
              Kernel: wimboot
              Name: windows-2012r2-install
              OS:
                Codename: ''
                Family: ''
                Files: null
                IsoFile: rebar-win2012r2.iso
                IsoSha256: 8ce9f9b23aaf397913d927c7fda3343d411f6846ef7b94f29b51108e95c46f6e
                IsoUrl: ''
                Name: windows-2012r2
                Version: ''
              RequiredParams:
                - logging_servers
                - ntp_servers
                - operating-system-disk
                - provisioner-default-password-hash
                - proxy-servers
                - use-proxy
                - rebar-access_keys
                - rebar-machine_key
                - operating-system-license-key
                - operating-system-install-flavor
              Templates:
                - Name: pxelinux
                  Path: 'pxelinux.cfg/{{.Machine.HexAddress}}'
                  UUID: windows.pxelinux.tmpl
                - Name: ipxe
                  Path: '{{.Machine.Address}}.ipxe'
                  UUID: windows.ipxe.tmpl
                - Name: start-install
                  Path: '{{.Machine.Path}}/stage1.ps1'
                  UUID: stage1.cmd.tmpl
                - Name: unattend.xml
                  Path: '{{.Machine.Path}}/unattend.xml'
                  UUID: 2012r2-unattend.xml.tmpl
              TenantId: 1
      responses:
        '200':
          description: ''
          schema:
            $ref: '#/definitions/bootenv-input'
        '401':
          $ref: '#/responses/trait:commonPutResponses:401'
        '404':
          $ref: '#/responses/trait:commonPutResponses:404'
        '500':
          $ref: '#/responses/trait:commonPutResponses:500'
    delete:
      operationId: delete-bootenv
      summary: Delete Bootenv
      tags:
        - Bootenvs
      responses:
        '204':
          description: ''
        '404':
          $ref: '#/responses/trait:commonDeleteResponses:404'
        '409':
          $ref: '#/responses/trait:commonDeleteResponses:409'
        '500':
          $ref: '#/responses/trait:commonDeleteResponses:500'
...
responses:
  'trait:commonGetResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonGetResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonGetResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:409':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonDeleteResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:409':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPostResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPutResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:401':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:404':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:417':
    description: ''
    schema:
      $ref: '#/definitions/result'
  'trait:commonPatchResponses:500':
    description: ''
    schema:
      $ref: '#/definitions/result'
definitions:
...
  result:
    title: Result
    type: object
    properties:
      Code:
        type: integer
      Messages:
        type: array
        items:
          type: string
...
  patchinput:
    title: PatchInput
    type: array
    items:
      $ref: '#/definitions/patchline'
  patchline:
    title: PatchLine
    type: object
    properties:
      Op:
        type: string
      Path:
        type: string
      From:
        type: string
      Value: {}
...
  bootenv-input:
    title: Bootenv Input
    type: object
    properties:
      Available:
        type: boolean
      BootParams:
        type: string
      Errors:
        type: array
        items:
          type: string
      Initrds:
        type: array
        items:
          type: string
      Kernel:
        type: string
      Name:
        type: string
      OS:
        type: object
        properties:
          Codename:
            type: string
          Family:
            type: string
          IsoFile:
            type: string
          IsoSha256:
            type: string
          IsoUrl:
            type: string
          Name:
            type: string
          Version:
            type: string
      RequiredParams:
        type: array
        items:
          type: string
      Templates:
        type: array
        items:
          type: object
          properties:
            Name:
              type: string
            Path:
              type: string
            UUID:
              type: string
    example:
      Available: false
      BootParams: ''
      Errors:
        - 'bootenv: windows-2012r2-install: missing kernel wimboot (/tftpboot/windows-2012r2/install/wimboot)'
        - 'bootenv: windows-2012r2-install: missing initrd boot/bcd (/tftpboot/windows-2012r2/install/boot/bcd)'
        - 'bootenv: windows-2012r2-install: missing initrd boot/boot.sdi (/tftpboot/windows-2012r2/install/boot/boot.sdi)'
        - 'bootenv: windows-2012r2-install: missing initrd rebar-winpe.wim (/tftpboot/windows-2012r2/install/rebar-winpe.wim)'
      Initrds:
        - boot/bcd
        - boot/boot.sdi
        - rebar-winpe.wim
      Kernel: wimboot
      Name: windows-2012r2-install
      OS:
        Codename: ''
        Family: ''
        Files: []
        IsoFile: rebar-win2012r2.iso
        IsoSha256: 8ce9f9b23aaf397913d927c7fda3343d411f6846ef7b94f29b51108e95c46f6e
        IsoUrl: ''
        Name: windows-2012r2
        Version: ''
      RequiredParams:
        - logging_servers
        - ntp_servers
        - operating-system-disk
        - provisioner-default-password-hash
        - proxy-servers
        - use-proxy
        - rebar-access_keys
        - rebar-machine_key
        - operating-system-license-key
        - operating-system-install-flavor
      Templates:
        - Name: pxelinux
          Path: 'pxelinux.cfg/{{.Machine.HexAddress}}'
          UUID: windows.pxelinux.tmpl
        - Name: ipxe
          Path: '{{.Machine.Address}}.ipxe'
          UUID: windows.ipxe.tmpl
        - Name: start-install
          Path: '{{.Machine.Path}}/stage1.ps1'
          UUID: stage1.cmd.tmpl
        - Name: unattend.xml
          Path: '{{.Machine.Path}}/unattend.xml'
          UUID: 2012r2-unattend.xml.tmpl
      TenantId: 1